+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Compilation - `CompilationNode`s take the full `CompilationConfig` as input (!615, #405)
 - Visualization - Introduce the `x_range` keyword for the matplotlib backend in `Schedule.plot_pulse_diagram`. This will cut off any points outside the given range when creating the plot. This can be used to reduce memory usage when plotting a small section of a long pulse sequence (!629).
 - Schedules - Add `nv_dark_esr_sched_nco` spectroscopy schedule using SetClockFrequency Operation to sweep the NCO frequency (!639)
+- ZI LabOne backend - Return datasets from UHFQA instrument coordinator component (which fixes the broken backend) (#410, !623).
 
 ## 0.12.3 (2023-03-09)